unifying clusterId to be uint16 (#2777)

+++ waku/waku_enr/sharding.nim
@@ -260,7 +260,7 @@ proc containsShard*(r: Record, topic: PubsubTopic | string): bool =
 
   containsShard(r, parseRes.value)
 
-proc isClusterMismatched*(record: Record, clusterId: uint32): bool =
+proc isClusterMismatched*(record: Record, clusterId: uint16): bool =
   ## Check the ENR sharding info for matching cluster id
   if (let typedRecord = record.toTyped(); typedRecord.isOk()):
     if (let relayShard = typedRecord.get().relaySharding(); relayShard.isSome()):
